long bug fixed

+++ core/src/main/java/site/ycsb/workloads/CoreWorkload.java
@@ -480,7 +480,7 @@ public class CoreWorkload extends Workload {
       // the keyspace doesn't change from the perspective of the scrambled zipfian generator
       final double insertproportion = Double.parseDouble(
           p.getProperty(INSERT_PROPORTION_PROPERTY, INSERT_PROPORTION_PROPERTY_DEFAULT));
-      int opcount = Integer.parseInt(p.getProperty(Client.OPERATION_COUNT_PROPERTY));
+      long opcount = Long.parseLong(p.getProperty(Client.OPERATION_COUNT_PROPERTY));
       int expectednewkeys = (int) ((opcount) * insertproportion * 2.0); // 2 is fudge factor
       double zipfianconstant = Double.parseDouble(
           p.getProperty("zipfian.constant", "0.99"));
@@ -499,7 +499,7 @@ public class CoreWorkload extends Workload {
     } else if (requestdistrib.compareTo("unscrambled_zipfian") == 0) {
       final double insertproportion = Double.parseDouble(
           p.getProperty(INSERT_PROPORTION_PROPERTY, INSERT_PROPORTION_PROPERTY_DEFAULT));
-      int opcount = Integer.parseInt(p.getProperty(Client.OPERATION_COUNT_PROPERTY));
+      long opcount = Long.parseLong(p.getProperty(Client.OPERATION_COUNT_PROPERTY));
       int expectednewkeys = (int) ((opcount) * insertproportion * 2.0); // 2 is fudge factor
       double zipfianconstant = Double.parseDouble(
           p.getProperty("zipfian.constant", "0.99"));
